Upload to quadtree system
This one makes it work a bit better, since it has a quadtree correction system. It happens so that if a ball falls on an edge of a sub-region, it will miss some of the possible collisions. This commit is supposed to change it, but it doesnt. Im just trying to keep everything cool on GitHub so I can learn faster. I will make it work, i promise!